requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.1
cssselect==1.0.3
fake-useragent==0.1.10
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
lxml==4.2.1
mmh3==2.5.1
mysqlclient==1.3.12
parsel==1.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
PyExecJS==1.5.1
pyOpenSSL==17.5.0
pywin32==223
queuelib==1.5.0
redis==2.10.6
requests==2.18.4
Scrapy==1.5.0
scrapy-proxies==0.4
scrapy-redis==0.6.8
selenium==3.11.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
urllib3==1.22
w3lib==1.19.0
zope.interface==4.4.3
